Improve histogram formatting functionality.
Rename ConsoleHistogramFormatter to
HistogramFormatter.

Add a HistogramFormat configuration class.

Move cucumberTest/ under src/ in the
stats-utils-measure-jsr363 project, because that's
where it should be and also because the Cucumber
task wasn't seeing it where it was (!).

Delete ConsoleHistogramQuantityFormatter because
at this point its functionality can literally be
replaced by a lambda (which is what
PowerPlantCapacityHistogramExample does now).

Improve padding of formatted percentage values.

Make the "All" label for a single-bucket histogram
(which is not expected to be used often) be a
configurable thing in the HistogramFormat.

Make percentage value formatting use a
NumberFormat, which was the last locale-specific
behavior of HistogramFormatter.

Add some tests for HistogramFormatter that verify
that its non-US-locale formatting works as
expected.

Remove ImmutableList from the API of
ImmutableHistogram, since Guava is not supposed to
be part of the public API of this library.